refactor(provider): move passport query port to application layer
