+++ backend/src/main/java/com/alligator/market/backend/provider/application/passport/catalog/PassportCatalogServiceImpl.java
@@ -2,7 +2,7 @@ package com.alligator.market.backend.provider.application.passport.catalog;
 
 import com.alligator.market.domain.provider.passport.ProviderPassport;
 import com.alligator.market.domain.provider.vo.ProviderCode;
-import com.alligator.market.domain.provider.readmodel.passport.query.port.ProviderPassportQueryPort;
+import com.alligator.market.backend.provider.application.passport.catalog.port.out.ProviderPassportQueryPort;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
